refactor(table): keep the refresh vocabulary out of the component files (LFE-14954)
structure:stats housekeeping on the files this branch touches, per the web
project-structure RFC (rule 2: a component file exports only the component
and its types):

- REFRESH_INTERVALS / RefreshInterval move from the refresh button to
  components/table/utils/refresh-intervals.ts, where the five importers now
  read them.
- shouldIgnoreRowClickTarget was exported from data-table.tsx but only ever
  used inside it; the internal helpers of useLiveTableDateRange were
  exported for no reader either. Both are module-local now.

rule 2: 55 -> 53, no other rule moved (total 2606 -> 2604).
