+++ web/src/components/table/data-table-refresh-button.tsx
@@ -9,16 +9,10 @@ import {
 } from "@/src/components/ui/dropdown-menu";
 import { cn } from "@/src/utils/tailwind";
 import { useAnimatedBusy } from "@/src/hooks/useAnimatedBusy";
-
-export const REFRESH_INTERVALS = [
-  { label: "Off", value: null },
-  { label: "30s", value: 30_000 },
-  { label: "1m", value: 60_000 },
-  { label: "5m", value: 300_000 },
-  { label: "15m", value: 900_000 },
-] as const;
-
-export type RefreshInterval = (typeof REFRESH_INTERVALS)[number]["value"];
+import {
+  REFRESH_INTERVALS,
+  type RefreshInterval,
+} from "@/src/components/table/utils/refresh-intervals";
 
 interface DataTableRefreshButtonProps {
   onRefresh: () => void;
