+++ web/src/components/table/table-header-controls.tsx
@@ -1,8 +1,6 @@
 import { TimeRangePicker } from "@/src/components/date-picker";
-import {
-  DataTableRefreshButton,
-  type RefreshInterval,
-} from "@/src/components/table/data-table-refresh-button";
+import { DataTableRefreshButton } from "@/src/components/table/data-table-refresh-button";
+import { type RefreshInterval } from "@/src/components/table/utils/refresh-intervals";
 import { PageHeaderControlsPortal } from "@/src/components/layouts/page-header-controls-slot";
 import {
   TABLE_AGGREGATION_OPTIONS,
