+++ web/src/components/table/utils/refresh-intervals.ts
@@ -1,0 +1,10 @@
+/** Auto-refresh choices offered by the table refresh control. */
+export const REFRESH_INTERVALS = [
+  { label: "Off", value: null },
+  { label: "30s", value: 30_000 },
+  { label: "1m", value: 60_000 },
+  { label: "5m", value: 300_000 },
+  { label: "15m", value: 900_000 },
+] as const;
+
+export type RefreshInterval = (typeof REFRESH_INTERVALS)[number]["value"];
